% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp2domains_by_twtest_parallel.R
\name{comp2domains_by_twtest_parallel}
\alias{comp2domains_by_twtest_parallel}
\title{compare multiple domains}
\usage{
comp2domains_by_twtest_parallel(
  fhic0,
  fhic1,
  tadlist_of_hic0.bed,
  ofile = paste(getwd(), "/stdout", sep = ""),
  min_nbin = 10,
  hicnorm = "KR",
  reso = 100000,
  f = 0.5
)
}
\arguments{
\item{fhic0}{the first hic file to be compared}

\item{fhic1}{the  second hic file to be compared}

\item{tadlist_of_hic0.bed}{domian list}

\item{ofile}{the file path of the result}

\item{min_nbin}{default is 10}

\item{hicnorm}{default is "KR"}

\item{reso}{resolution to choose,default is 100000}

\item{f}{default is 0.5}
}
\value{
result of single pair comparition
}
\description{
compare multiple domains
}
\examples{
comp2domains_by_twtest_parallel(
'/lanec1_home/data/GSE63525_cell2014/GSE63525_GM12878_insitu_primary_replicate_combined.hic',
'/lanec1_home/data/GSE63525_cell2014/GSE63525_KBM7_combined.hic',
'/lanec1_home/data/tads/ArrowheadDomainlist/GSE63525_GM12878_primary+replicate_Arrowhead_domainlist.txt',
ofile = '/lanec2_home/zhangx/DiffCompare/pythonR_compare_result/GM12878_KBM7_stdout_R')
}
