% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix_from_hic.R
\name{contact_matrix_from_hic}
\alias{contact_matrix_from_hic}
\title{Return matrix from hic file}
\usage{
contact_matrix_from_hic(chrn, start, end, reso, fhic, hicnorm)
}
\arguments{
\item{chrn}{chosen chromosome}

\item{start}{start bin of chromosome}

\item{end}{end bin of chromosome}

\item{reso}{chosen resoltion}

\item{fhic}{chosen hic file}

\item{hicnorm}{default is "KR"}
}
\value{
matrix
}
\description{
Return matrix from hic file
}
\examples{
contact_matrix_from_hic(1,1060000,1080000,10000,'/lanec1_home/data/GSE63525_cell2014/K562_combined.hic','KR')
}
