% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makewindow2.R
\name{makewindow2}
\alias{makewindow2}
\title{Return windows between start and end,based on the resolution}
\usage{
makewindow2(start, end, reso)
}
\arguments{
\item{start}{start bin of chosen chromosome}

\item{end}{end bin of chosen chromosome}

\item{reso}{chosen resolution}
}
\value{
wins
}
\description{
Return windows between start and end,based on the resolution
}
\examples{
makewindow2(1060000,1080000,10000)
}
