% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDiffbyMeanSD.R
\name{normDiffbyMeanSD}
\alias{normDiffbyMeanSD}
\title{Return nomalization matrix based on its mean and standard deviation}
\usage{
normDiffbyMeanSD(D)
}
\arguments{
\item{D}{matrix}
}
\value{
normalized matrix
}
\description{
Return nomalization matrix based on its mean and standard deviation
}
\examples{
normDiffbyMeanSD(matrix)
}
