% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdAdjust.R
\name{sdAdjust}
\alias{sdAdjust}
\title{Return adjust standard deviation}
\usage{
sdAdjust(numberin)
}
\arguments{
\item{numberin}{number to be calculated}
}
\value{
adjusted standard deviation
}
\description{
Return adjust standard deviation
}
\examples{
sdAdjust(c(1,2,3,4,5))
}
