% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twtest_formula.R
\name{twtest_formula}
\alias{twtest_formula}
\title{Return twtest result of matrix}
\usage{
twtest_formula(Mat)
}
\arguments{
\item{Mat}{input matrix}
}
\value{
test result of chosen matrix
}
\description{
Return twtest result of matrix
}
\examples{
twtest_formula(matrix
}
